changed to normalize RDF by 4/3PI(r2^3-r1^3)*numFeatures/TotalBoxVolume. Removed normalization in BC calculation.

+++ Source/Plugins/Statistics/StatisticsFilters/GenerateEnsembleStatistics.cpp
@@ -375,7 +375,7 @@ void GenerateEnsembleStatistics::dataCheck()
   if(m_IncludeRadialDistFunc == true)
   {
    // dims[0]=10;
-    DataArray<int>::Pointer tempPtr = getDataContainerArray()->getExistingPrereqArrayFromPath<DataArray<int>, AbstractFilter>(this, getRDFArrayPath());
+    DataArray<float>::Pointer tempPtr = getDataContainerArray()->getExistingPrereqArrayFromPath<DataArray<float>, AbstractFilter>(this, getRDFArrayPath());
     if (NULL != tempPtr.get())
     {
         m_RadialDistFuncPtr = tempPtr;
